Refactor Output as System.console() is not applicable
